completed the work ;}

+++ controller/contactconttroller.js
@@ -2,14 +2,14 @@ const asynchandler = require("express-async-handler")
 const Contact = require("../models/contactmodel")
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asynchandler(async(req,res)=>{
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts);
 });
 //@desc create contacts
 //@route POST /api/contacts
-//@access public
+//@access private
 const createcontact = asynchandler(async(req,res)=>{
     console.log("the req body is :", req.body)
     const {email,name,phone}=req.body
@@ -17,12 +17,12 @@ const createcontact = asynchandler(async(req,res)=>{
         res.status(400);
         throw new Error("all file manditory");
     }
-    const contact = await Contact.create({email,name,phone});
+    const contact = await Contact.create({email,name,phone,user_id:req.user.id});
     res.status(201).json(contact);
 });
 //@desc Get all contacts
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asynchandler(async(req,res)=>{
     const contacts = await Contact.findById(req.params.id);
     if(!contacts){
@@ -33,12 +33,16 @@ const getContact = asynchandler(async(req,res)=>{
 });
 //@desc update contacts
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asynchandler(async(req,res)=>{
     const contacts = await Contact.findById(req.params.id);
     if(!contacts){
         res.status(404);
         throw new Error("contact not found")
+    }
+    if(contacts.user_id.toString() != req.user.id){
+        res.status(403);
+        throw new error("other user contact"); 
     }
     const updatecontact = await Contact.findByIdAndUpdate(
         req.params.id,
@@ -49,7 +53,7 @@ const updateContact = asynchandler(async(req,res)=>{
 });
 //@desc delete contacts
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asynchandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
 
@@ -57,10 +61,13 @@ const deleteContact = asynchandler(async (req, res) => {
         res.status(404);
         throw new Error("Contact not found");
     }
+    if(contact.user_id.toString() != req.user.id){
+        res.status(403);
+        throw new error("other user contact"); 
+    }
+    await contact.deleteOne();
 
-    await contact.deleteOne(); // Proper deletion
-
-    res.status(200).json({ message: "Contact deleted successfully" });
+    res.status(200).json(contact);
 });
 module.exports={
     getContacts,createcontact,updateContact,deleteContact,getContact
